move storage keys to xcm agent
