+++ packages/server/src/services/ingress/watcher/local-cache.ts
@@ -9,9 +9,9 @@ import type { SignedBlockExtended } from '@polkadot/api-derive/types'
 import type { Raw } from '@polkadot/types'
 import type { Hash } from '@polkadot/types/interfaces'
 
+import { parachainSystemHrmpOutboundMessages, parachainSystemUpwardMessages } from '../../agents/xcm/storage.js'
 import { NetworkConfiguration } from '../../config.js'
 import { Janitor } from '../../persistence/janitor.js'
-import { parachainSystemHrmpOutboundMessages, parachainSystemUpwardMessages } from '../../subscriptions/storage.js'
 import { HexString } from '../../subscriptions/types.js'
 import { TelemetryEventEmitter } from '../../telemetry/types.js'
 import { DB, Logger, NetworkURN, Services, prefixes } from '../../types.js'
